refactor: extract briefToHtml into lib/morning-brief-html for testability
morning-brief-cron transitively imports lib/redis (via app/actions/weather), which throws at module load when REDIS_URL is unset — as it is in the unit-tests CI env. Move the pure HTML rendering helpers (escapeHtml, formatStaffLine, briefToHtml) into a dependency-free module so the whitelabelling test can import them without pulling in redis.
